Changed context::ShutterOpen/Close return type from float to double.
This is more consistent with the representation of time in both Houdini and NSI,
and should help avoid numerical precision issues (by avoiding conversions from
double to float, and then back to double).

+++ context.h
@@ -54,9 +54,9 @@ public:
 	/// Returns true if motion blur is required for this render
 	bool MotionBlur()const { return m_shutter > 0.0f; }
 	/// Returns the time at which the shutter opens
-	float ShutterOpen()const { return m_current_time - m_shutter/2.0f; }
+	double ShutterOpen()const { return m_current_time - m_shutter/2.0f; }
 	/// Returns the time at which the shutter closes
-	float ShutterClose()const { return m_current_time + m_shutter/2.0f; }
+	double ShutterClose()const { return m_current_time + m_shutter/2.0f; }
 
 	/// Returns true if a single frame is to be rendered
 	bool SingleFrame()const { return m_start_time == m_end_time; }
